Move exec.HostRunner to rig.Executor
